# docs/swagger.yaml
basePath: /api
definitions:
...
  domain.Package:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      latest_release:
        $ref: '#/definitions/domain.Release'
      name:
        type: string
      project_id:
        type: string
      release_count:
        type: integer
      total_downloads:
        type: integer
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  domain.Release:
    properties:
      changelog:
        description: Release notes/changelog
        type: string
      created_at:
        type: string
      created_by:
        type: string
      download_count:
        type: integer
      file_hash:
        type: string
      file_name:
        type: string
      file_path:
        type: string
      file_size:
        type: integer
      id:
        type: string
      package_id:
        type: string
      share_expiry:
        type: string
      share_token:
        type: string
      tag_name:
        type: string
      version_code:
        type: string
      version_name:
        type: string
    type: object
...
  domain.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Package Management System API for managing projects, packages, and
    users with RBAC.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PKMS API
  version: "1.0"
paths:
...
  /packages:
    get:
      consumes:
      - application/json
      description: Get all packages with pagination and optional project filtering
      parameters:
      - description: Tenant ID
        in: header
        name: x-tenant-id
        required: true
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 20)'
        in: query
        name: pageSize
        type: integer
      - description: Filter by project ID
        in: query
        name: project_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved packages
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.Package'
                  type: array
              type: object
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/domain.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Get all packages
      tags:
      - Packages
    post:
      consumes:
      - application/json
      description: Create a new package with provided details
      parameters:
      - description: Tenant ID
        in: header
        name: x-tenant-id
        required: true
        type: string
      - description: Package details
        in: body
        name: package
        required: true
        schema:
          $ref: '#/definitions/domain.Package'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created package
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Package'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/domain.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Create a new package
      tags:
      - Packages
  /packages/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific package by its ID
      parameters:
      - description: Tenant ID
        in: header
        name: x-tenant-id
        required: true
        type: string
      - description: Package ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Package deleted successfully
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Invalid package ID
          schema:
            $ref: '#/definitions/domain.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.Response'
        "404":
          description: Package not found
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Delete package by ID
      tags:
      - Packages
    get:
      consumes:
      - application/json
      description: Get a specific package by its ID
      parameters:
      - description: Tenant ID
        in: header
        name: x-tenant-id
        required: true
        type: string
      - description: Package ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved package
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Package'
              type: object
        "400":
          description: Invalid package ID
          schema:
            $ref: '#/definitions/domain.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.Response'
        "404":
          description: Package not found
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Get package by ID
      tags:
      - Packages
    put:
      consumes:
      - application/json
      description: Update a specific package with new details
      parameters:
      - description: Tenant ID
        in: header
        name: x-tenant-id
        required: true
        type: string
      - description: Package ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated package details
        in: body
        name: package
        required: true
        schema:
          $ref: '#/definitions/domain.Package'
      produces:
      - application/json
      responses:
        "200":
          description: Package updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Invalid request body or package ID
          schema:
            $ref: '#/definitions/domain.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.Response'
        "404":
          description: Package not found
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Update package by ID
      tags:
      - Packages
...
  /projects/{id}/packages:
    get:
      consumes:
      - application/json
      description: Retrieve all packages associated with a specific project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Project packages retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items: {}
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Get project packages
      tags:
      - Projects
  /projects/{projectId}/packages:
    get:
      consumes:
      - application/json
      description: Get all packages belonging to a specific project
      parameters:
      - description: Tenant ID
        in: header
        name: x-tenant-id
        required: true
        type: string
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved project packages
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.Package'
                  type: array
              type: object
        "400":
          description: Invalid project ID
          schema:
            $ref: '#/definitions/domain.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.Response'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Get all packages for a project
      tags:
      - Packages
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"